fixed Andrey's revisions
