+++ src/app/page.tsx
@@ -3,12 +3,12 @@ import { Footer } from "@/shared/ui/footer";
 
 export default function Home() {
     return (
-        <section className="">
+        <section>
             <div className="h-full overflow-y-scroll grid grid-cols-2">
                 <div className="flex items-center justify-center">
                     <p className="text-8xl tracking-tighter text-chart-5 font-black ">TENNIS<br />STORE</p>
                 </div>
-                <div className="">
+                <div>
                     <CarouselWrapper />
                 </div>
             </div>
